feat(common): responsive modal window layout, no bulma
