+++ components/common/modal/modal.scss
@@ -1,0 +1,78 @@
+@import 'responsiveness';
+@import 'variables';
+
+.wir-modal {
+  height: 100%;
+  left: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+
+  &__background {
+    background-color: $primary-black;
+    height: 100%;
+    opacity: 0.5;
+    position: fixed;
+    width: 100%;
+    z-index: 100; // same with .wir-overlay
+  }
+
+  &__content {
+    background-color: $primary-white;
+    position: absolute;
+    z-index: 200; // same with .wir-sidebar
+  }
+
+  &__close {
+    color: $primary;
+    cursor: pointer;
+    font-size: 20px;
+    position: absolute;
+    right: 20px;
+    top: 20px;
+  }
+}
+
+@mixin screen-desktop-tablet-large {
+  .wir-modal {
+    &__content {
+      margin: 70px calc((100vw - 690px) / 3) 0 calc((100vw - 690px) / 6);
+      padding: 20px calc((100vw - 690px) / 3) 90px calc((100vw - 690px) / 6);
+    }
+  }
+}
+
+@include screen-desktop {
+  @include screen-desktop-tablet-large;
+}
+
+@include screen-tablet-large {
+  @include screen-desktop-tablet-large;
+}
+
+@include screen-tablet {
+  .wir-modal {
+    &__content {
+      margin: 70px calc((100vw - 530px) / 3) 0 calc((100vw - 530px) / 6);
+      padding: 20px calc((100vw - 530px) / 3) 90px calc((100vw - 530px) / 6);
+    }
+  }
+}
+
+@include screen-touch {
+  .wir-modal {
+    &__content {
+      margin: 70px calc((100vw - 450px) / 3) 0 calc((100vw - 450px) / 6);
+      padding: 20px calc((100vw - 450px) / 3) 90px calc((100vw - 450px) / 6);
+    }
+  }
+}
+
+@include screen-mobile {
+  .wir-modal {
+    &__content {
+      margin: 0;
+      padding: 20px calc((100vw - 302px) * 2 / 3) 0 calc((100vw - 302px) / 3);
+    }
+  }
+}
